requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
python-dotenv==1.2.1
cryptography==46.0.3